Source/SDK/PayPal/Api/Payments/RedirectUrls.cs
using System;
using Newtonsoft.Json;
using Newtonsoft.Json.Serialization;
using PayPal.Api.Validation;

namespace PayPal.Api.Payments
{
    public class RedirectUrls
    {
        /// <summary>
        /// Url where the payer would be redirected to after approving the payment.
        /// </summary>
        [JsonProperty(NullValueHandling = NullValueHandling.Ignore)]
        public string return_url { get; set; }

        /// <summary>
        /// Url where the payer would be redirected to after canceling the payment.
        /// </summary>
        [JsonProperty(NullValueHandling = NullValueHandling.Ignore)]
        public string cancel_url { get; set; }

        /// <summary>
        /// Converts the object to JSON string
        /// </summary>
        public virtual string ConvertToJson()
        {
            return JsonFormatter.ConvertToJson(this);
        }
    }
}
